requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
scikit-learn>=1.0.0
joblib>=1.1.0
pyyaml>=5.4.1
pydantic>=1.8.0
shap>=0.40.0
streamlit>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0